+++ LookupAnything/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed ingredient names in recipes that need any egg or any milk.
+* Improved controller support in the search menu.
 
 ## 1.38.2
 Released 09 January 2023 for SMAPI 3.14.0 or later.